+++ src/refactoring/components/AdminPage.tsx
@@ -1,7 +1,7 @@
 import { useState } from "react";
 import { Coupon, Product } from "../../types.ts";
 import { useProductsContext } from "../hooks";
-import { Products } from "./Products.tsx";
+import { Products } from "./AdminProducts.tsx";
 
 interface Props {
   coupons: Coupon[];
