+++ src/refactoring/components/AdminProducts.tsx
